Batch refactored

+++ backend/explorer-core/autoconfigure/src/main/java/org/royllo/explorer/core/provider/tapd/ProofRequest.java
@@ -26,7 +26,7 @@ public class ProofRequest {
      * The index depth of the decoded proof, with 0 being the latest proof.
      */
     @JsonProperty("proof_at_depth")
-    int proofAtDepth;
+    long proofAtDepth;
 
     /**
      * With prev witnesses.
